requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.6
dj-database-url==1.0.0
Django==4.1.3
django-environ==0.9.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.4
gunicorn==20.1.0
iniconfig==1.1.1
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.2.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.6
whitenoise==6.2.0
